refactored required fields validation
