+++ EventsExpress/ClientApp/src/components/helpers/validators/email-address-validator.js
@@ -1,0 +1,9 @@
+export const isValidEmail = email => {
+    const errors = {};
+    if (email &&
+        !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(email)) {
+        errors.email = 'Invalid email address'
+    }
+    return errors;
+}
+
