--- EventsExpress/ClientApp/src/components/helpers/validators/email-field-validator.js
@@ -1,13 +1,0 @@
-export const emailField = values => {
-    const errors = {};
-
-    if (!values.email) {
-        errors.email = 'Required'
-    }
-    if (values.email &&
-        !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
-        errors.email = 'Invalid email address'
-    }
-    return errors;
-}
-
