+++ EventsExpress/ClientApp/src/components/recoverPassword/recoverPassword.js
@@ -5,10 +5,14 @@ import Button from "@material-ui/core/Button";
 import DialogContentText from '@material-ui/core/DialogContentText';
 import ErrorMessages from '../shared/errorMessage';
 import { renderTextField } from '../helpers/form-helpers';
-import { emailField } from '../helpers/validators/email-field-validator';
+import { isValidEmail } from '../helpers/validators/email-address-validator';
+import { fieldIsRequired } from '../helpers/validators/required-fields-validator';
 
 const validate = values => {
-    return emailField(values)
+    return {
+        ...fieldIsRequired(values, requiredFields),
+        ...isValidEmail(values.email)
+    }
 }
 
 class RecoverPassword extends React.Component {
