+++ EventsExpress/ClientApp/src/containers/event-edit-validate-form .js
@@ -1,18 +1,20 @@
 import { numberField } from '../components/helpers/validators/number-fields-validator';
-import { requiredField } from '../components/helpers/validators/required-fields-validator';
+import { fieldIsRequired } from '../components/helpers/validators/required-fields-validator';
 
 export const validate = values => {
     let errors = {};
     const occurenceFields = ['periodicity', 'frequency'];
+    const requiredFields = [
+        'title',
+        'description',
+        'categories',
+        'image',
+    ];
     occurenceFields.forEach(field => {
         if ('occurenceFields'.checked && !values[field]) {
             errors[field] = 'Required'
         }
     })
-
-    var requiredErrors = requiredField(values);
-    var numberErrors = numberField(values);
-    errors = { ...errors, ...numberErrors, ...requiredErrors };
 
     if (values.categories != null && values.categories.length == 0) {
         errors.categories = "Required";
@@ -25,5 +27,9 @@ export const validate = values => {
     }
     if (values.image != null && values.image.file != null && values.image.file.size < 4096)
         errors.image = "Image is too small";
-    return errors;
+    return {
+        ...errors,
+        ...numberField(values),
+        ...fieldIsRequired(values, requiredFields),
+    }
 }
